:recycle: Tidy up context creation and browser utils
