+++ src/core/context.js
@@ -1,43 +1,30 @@
+import dummy from './utils/dummy';
 import FakeContext from './utils/fake-context';
+import iOS from './utils/iOS';
 
-function getContext() {
-    const desiredSampleRate = 44100;
+const desiredSampleRate = 44100;
 
-    const Ctx = window.AudioContext || window.webkitAudioContext || FakeContext;
+const Ctx = window.AudioContext || window.webkitAudioContext || FakeContext;
 
-    let ctx = new Ctx();
+let context = new Ctx();
 
-    if (!ctx) {
-        ctx = new FakeContext();
-    }
-
-	// Check if hack is necessary. Only occurs in iOS6+ devices
-	// and only when you first boot the iPhone, or play a audio/video
-	// with a different sample rate
-	// https://github.com/Jam3/ios-safe-audio-context/blob/master/index.js
-    if (/(iPhone|iPad)/i.test(navigator.userAgent) && ctx.sampleRate !== desiredSampleRate) {
-        const buffer = ctx.createBuffer(1, 1, desiredSampleRate);
-        const dummy = ctx.createBufferSource();
-        dummy.buffer = buffer;
-        dummy.connect(ctx.destination);
-        dummy.start(0);
-        dummy.disconnect();
-
-        ctx.close(); // dispose old context
-        ctx = new Ctx();
-    }
-
-	// Handles bug in Safari 9 OSX where AudioContext instance starts in 'suspended' state
-
-    const isSuspended = ctx.state === 'suspended';
-
-    if (isSuspended && typeof ctx.resume === 'function') {
-        window.setTimeout(function() {
-            ctx.resume();
-        }, 1000);
-    }
-
-    return ctx;
+if (!context) {
+    context = new FakeContext();
 }
 
-export default getContext();
+// Check if hack is necessary. Only occurs in iOS6+ devices
+// and only when you first boot the iPhone, or play a audio/video
+// with a different sample rate
+// https://github.com/Jam3/ios-safe-audio-context/blob/master/index.js
+if (iOS && context.sampleRate !== desiredSampleRate) {
+    dummy(context);
+    context.close(); // dispose old context
+    context = new Ctx();
+}
+
+// Handles bug in Safari 9 OSX where AudioContext instance starts in 'suspended' state
+if (context.state === 'suspended' && typeof context.resume === 'function') {
+    window.setTimeout(() => context.resume(), 1000);
+}
+
+export default context;
