+++ src/core/utils/dummy.js
@@ -1,0 +1,9 @@
+export default function dummy(context) {
+    const buffer = context.createBuffer(1, 1, context.sampleRate);
+    const source = context.createBufferSource();
+    source.buffer = buffer;
+    source.connect(context.destination);
+    source.start(0);
+    source.stop(0);
+    source.disconnect();
+}
