+++ src/core/utils/iOS.js
@@ -1,0 +1,1 @@
+export default navigator && /(iPhone|iPad|iPod)/i.test(navigator.userAgent);
