+++ src/core/utils/pageVisibility.js
@@ -1,0 +1,30 @@
+export default function pageVisibility(onHidden, onShown) {
+    let hidden = null;
+    let visibilityChange = null;
+
+    if (typeof document.hidden !== 'undefined') {
+        hidden = 'hidden';
+        visibilityChange = 'visibilitychange';
+    } else if (typeof document.mozHidden !== 'undefined') {
+        hidden = 'mozHidden';
+        visibilityChange = 'mozvisibilitychange';
+    } else if (typeof document.msHidden !== 'undefined') {
+        hidden = 'msHidden';
+        visibilityChange = 'msvisibilitychange';
+    } else if (typeof document.webkitHidden !== 'undefined') {
+        hidden = 'webkitHidden';
+        visibilityChange = 'webkitvisibilitychange';
+    }
+
+    function onChange() {
+        if (document[hidden]) {
+            onHidden();
+        } else {
+            onShown();
+        }
+    }
+
+    if (typeof visibilityChange !== 'undefined') {
+        document.addEventListener(visibilityChange, onChange, false);
+    }
+}
